requirements.txt
rich
psutil
questionary
speedtest-cli
requests
GPUtil
py-cpuinfo
plyer
wmi
pynput
